requirements.txt
functions-framework
requests
google-genai
google-cloud-bigquery
google-cloud-tasks
google-cloud-secret-manager
google-cloud-firestore